Here we go.. getting there ;)
I re-read/skimmed the last month of alfs-discuss while writting this part
up.  So i've tried to take everyone's opinion on the best syntax when I
wrote this.  I know many will speak up when this get's checked in a few
sec, please do.. i always enjoy the comments.  And this isn't "set in
stone", not yet.

But it will be.  Fairly soon, for 1.0.  I know i'm not the only one that's
getting sick and tired of talking about this syntax over and over again *g*