Should have commited this a while ago, been a little lazy, considering nobody has asked me much about halfling recently ;-)
By default however, you will have to uncomment the code in execute()
located in lib.c, for it to actually do something.  I have not yet
implemented <stage>.  So, instead i've added in <setenv> support
for now, untill i've gotten around to it.  And because of that, there
is no chroot support, either.

Given that, it does still work.  I have tested it with a set of LFS
CVS profiles and a few extra things things from the blfs-book.  You
can do a full LFS build using it as-is, by manually chroot'ing, etc.

So.. you *can* use it as-is, however I wont recomment it, there is
hardly any comments in the code, neither is there any documentation
or much (or any ;p) error checking.

Anyways, enjoy!  More commits coming soon...